shift accounts views to bespoke components
