+++ src/components/accounts/AccountTable.tsx
@@ -1,0 +1,139 @@
+import { Pencil, Search, Trash2 } from 'lucide-react';
+import React, { useState } from 'react';
+import { type ColumnProps, type Key, Group } from 'react-aria-components';
+import { NavigateFunction } from 'react-router-dom';
+import type { Dispatch } from 'redux';
+import type { AnyAction } from 'starfx';
+
+import { type Account } from '~/src/store/schema.ts';
+import { accountRemove } from '~/src/store/thunks/accounts.ts';
+import {
+  floatFromDinero,
+  floatFromScaled,
+  toHumanCurrency,
+  toHumanInterest
+} from '~/src/store/utils/dineroUtils.ts';
+
+import {
+  Cell,
+  Column,
+  Row,
+  singleSort,
+  Table,
+  TableBody,
+  TableHeader
+} from '~/src/components/Table';
+
+import { Button } from '~/src/elements/Button.tsx';
+
+const navigateToAccountForm = (account: Account) => {
+  return {
+    state: {
+      account: {
+        id: account.id,
+        name: account.name,
+        starting: floatFromDinero(account.starting),
+        interest: floatFromScaled(account.interest),
+        vehicle: account.vehicle
+      }
+    }
+  };
+};
+
+export const AccountTable = ({
+  label,
+  accounts,
+  navigate,
+  dispatch
+}: {
+  label: string;
+  accounts: Account[];
+  navigate: NavigateFunction;
+  dispatch: Dispatch<AnyAction>;
+}) => {
+  const [sortable, setSetSortable] = useState<{
+    column: Key;
+    direction: 'ascending' | 'descending';
+  }>({ column: 'Vehicle', direction: 'descending' });
+
+  return (
+    <Table
+      aria-label={label}
+      selectionMode="none"
+      onSortChange={(item) => setSetSortable(item)}
+      sortDescriptor={sortable}
+    >
+      <TableHeader>
+        <Column id="name" defaultWidth="3fr" isRowHeader allowsSorting>
+          Name
+        </Column>
+        {(
+          [
+            ['Starting', '2fr'],
+            ['Interest', '1fr'],
+            ['Vehicle', '2fr'],
+            ['Actions', '2fr']
+          ] as [string, ColumnProps['defaultWidth']][]
+        ).map((h) => (
+          <Column key={h[0]} id={h[0]} defaultWidth={h[1]} allowsSorting>
+            {h[0]}
+          </Column>
+        ))}
+      </TableHeader>
+      <TableBody renderEmptyState={() => 'No accounts.'}>
+        {accounts.sort(singleSort(sortable)).map((account) => (
+          <AccountRow
+            key={account.id}
+            account={account}
+            navigate={navigate}
+            dispatch={dispatch}
+          />
+        ))}
+      </TableBody>
+    </Table>
+  );
+};
+
+const AccountRow = ({
+  account,
+  navigate,
+  dispatch
+}: {
+  account: Account;
+  navigate: NavigateFunction;
+  dispatch: Dispatch<AnyAction>;
+}) => (
+  <Row>
+    <Cell>{account.name}</Cell>
+    <Cell>{toHumanCurrency(account.starting)}</Cell>
+    <Cell>{toHumanInterest({ ...account.interest, trailingSymbol: '%' })}</Cell>
+    <Cell>{account.vehicle}</Cell>
+    <Cell>
+      <Group aria-label="Actions" className="space-x-1">
+        <Button
+          aria-label="View"
+          onPress={() => navigate(`/accounts/view/${account.id}`)}
+          className="px-0.5"
+        >
+          <Search className="max-h-3" />
+        </Button>
+        <Button
+          aria-label="Modify"
+          onPress={() =>
+            navigate('/accounts/set', navigateToAccountForm(account))
+          }
+          className="px-0.5"
+        >
+          <Pencil className="max-h-3" />
+        </Button>
+        <Button
+          aria-label="Delete"
+          onPress={() => dispatch(accountRemove({ id: account.id }))}
+          className="px-0.5"
+        >
+          <Trash2 className="max-h-3" />
+        </Button>
+      </Group>
+    </Cell>
+  </Row>
+);
