+++ src/pages/accounts/index.tsx
@@ -1,13 +1,17 @@
 import React from 'react';
 
-import Accounts from '../planning/accounts.tsx';
+import Accounts from '../../components/accounts/index.tsx';
 import BarChart from './barChart';
+import { schema } from '~/src/store/schema.ts';
+import { useSelector } from 'starfx/react';
 
 const AccountOverview = () => {
+  const accounts = useSelector(schema.accounts.selectTableAsList);
+
   return (
     <>
       <h1 className="text-3xl font-semibold">Accounts</h1>
-      <Accounts />
+      <Accounts accounts={accounts} />
       {/* <div className="my-2 py-1 overflow-hidden shadow rounded-lg divide-y divide-gray-200">
         <div className="px-4 py-5 sm:px-6">
           <h2 className="text-2xl font-semibold">{account.account.name}</h2>
